Refactor testing helper functions to rgb/contract/mod.rs
