+++ src/rgb/contract/amount.rs
@@ -443,11 +443,9 @@ mod serde_impl {
 
 #[cfg(test)]
 mod test {
-    use crate::paradigms::client_side_validation::{CommitEncode, Conceal};
-    use crate::strict_encoding::{Error, StrictDecode, StrictEncode};
-    use std::fmt::Debug;
-
+    use crate::strict_encoding::StrictDecode;
     use super::*;
+    use super::super::test_helpers::*;
 
     static AMOUNT: [u8; 43] = [
         0x3, 0x41, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x20, 0x0, 0xa6, 0x2b, 0x27, 0xae, 0x5a, 0xf,
@@ -504,94 +502,16 @@ mod test {
         0xea, 0x35, 0x59, 0x4a, 0xb6, 0xd1, 0x8e, 0x72, 0x0, 0x12, 0x67, 0xbf, 0xe, 0x42, 0x93,
         0xbf, 0x1d, 0x10, 0x75, 0xc0, 0xf6, 0x9c,
     ];
-    // Test suite function to test against the vectors
-    fn test_suite<T: StrictEncode + StrictDecode + PartialEq + Debug>(
-        object: &T,
-        test_vec: &[u8],
-        test_size: usize,
-    ) -> Result<T, Error> {
-        let mut encoded_object: Vec<u8> = vec![];
-        let write_1 = object.strict_encode(&mut encoded_object).unwrap();
-        let decoded_object = T::strict_decode(&encoded_object[..]).unwrap();
-        assert_eq!(write_1, test_size);
-        assert_eq!(decoded_object, *object);
-        encoded_object.clear();
-        let write_2 = decoded_object.strict_encode(&mut encoded_object).unwrap();
-        assert_eq!(encoded_object, test_vec);
-        assert_eq!(write_2, test_size);
-        Ok(decoded_object)
-    }
-
-    // Macro to run test_suite
-    macro_rules! test_encode {
-        ($($x:ident),*) => (
-            {
-                $(
-                    let object = Revealed::strict_decode(&$x[..]).unwrap();
-                    assert!(test_suite(&object, &$x[..], $x.to_vec().len()).is_ok());
-                )*
-            }
-        );
-    }
-
-    // Macro to run test suite with garbage vector
-    // Should produce "EnumValueNotKnown" error
-    macro_rules! test_garbage {
-        ($($x:ident),*) => (
-            {
-                $(
-                    let mut cp = $x.clone();
-                    cp[0] = 0x36 as u8;
-                    assert!(Revealed::strict_decode(&cp[..]).is_err());
-                )*
-            }
-        );
-    }
-
-    fn test_confidential<T>(data: &[u8], commitment: &[u8]) -> Result<T, Error>
-    where
-        T: Conceal + StrictDecode + StrictEncode + Clone + CommitEncode,
-        <T as Conceal>::Confidential: StrictDecode + StrictEncode + Eq,
-    {
-        // Create the Revealed Structure from data bytes
-        let revealed = T::strict_decode(data).unwrap();
-
-        // Conceal the Revealed structure into Confidential
-        let confidential = revealed.conceal();
-
-        // Strict_encode Confidential data
-        let mut confidential_encoded = vec![];
-        confidential
-            .strict_encode(&mut confidential_encoded)
-            .unwrap();
-
-        // strict_encode Revealed data
-        let mut revealed_encoded: Vec<u8> = vec![];
-        revealed.strict_encode(&mut revealed_encoded).unwrap();
-
-        // Assert encoded Confidential matches precomputed vector
-        assert_eq!(commitment, confidential_encoded);
-
-        // Assert encoded Confidential and Revealed are not equal
-        assert_ne!(confidential_encoded.to_vec(), revealed_encoded);
-
-        // commit_encode Revealed structure
-        let mut commit_encoded_revealed = vec![];
-        revealed.clone().commit_encode(&mut commit_encoded_revealed);
-
-        // Assert commit_encode and encoded Confidential matches
-        assert_eq!(commit_encoded_revealed, confidential_encoded);
-
-        // Assert commit_encode and precomputed Confidential matches
-        assert_eq!(commit_encoded_revealed, commitment);
-
-        Ok(revealed)
-    }
 
     #[test]
     fn test_amount() {
         test_encode!(AMOUNT);
-        test_garbage!(AMOUNT);
         assert!(test_confidential::<Revealed>(&AMOUNT, &CONFIDENTIAL_AMOUNT).is_ok());
+    }
+
+    #[test]
+    #[should_panic(expected = "EnumValueNotKnown")]
+    fn test_garbage() {
+        test_garbage!(AMOUNT);
     }
 }
